A light sensitive LED/SMD LED
This is arduino UNO build with two LEDs (one SMD, one regular) using pwm pins 9, 10, 11 for the RGB (not respective, check the code for actual mapping of RGB pins) control. There is a photo-resistor in this circuit as well, mounted on the opposite end of a full size breadboard from the LEDs to minimize the light contaminating the reading. The photo resistor is connected to the A0 pin on the arduino.